requirement.txt
annotated-types==0.7.0
anyio==4.5.2
arabic-reshaper==2.1.1
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
backports.zoneinfo @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_29zz4uumbw/croots/recipe/backports.zoneinfo_1661287541232/work
bcrypt==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.1.8
colorama==0.4.6
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==4.2.19
django-autoslug==1.9.8
django-bootstrap5==24.3
django-ckeditor==6.0.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
future==0.18.2
greenlet==3.1.1
gunicorn==20.0.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==2.10
imageio==2.14.0
Jinja2==3.1.5
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
networkx==2.6.3
Pillow==8.0.1
psycopg2 @ file:///C:/b/abs_4aotb39ojr/croot/psycopg2_1704232487763/work
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==1.7.1
pymongo==3.11.4
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2020.4
PyWavelets==1.2.0
PyYAML==6.0.2
razorpay==1.2.0
reportlab==3.5.56
requests==2.25.1
rich==13.9.4
rich-toolkit==0.13.2
scikit-image==0.19.1
shellingham==1.5.4
six==1.15.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlalchemy-sqlitecloud==0.1.2
sqlitecloud==0.0.83
sqlparse @ file:///C:/b/abs_88361ub_qu/croot/sqlparse_1690904577514/work
starlette==0.44.0
tifffile==2021.11.2
twilio==6.51.1
typer==0.15.1
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2022.1
urllib3==1.26.2
uvicorn==0.33.0
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
whitenoise==5.1.0
wincertstore==0.2
xhtml2pdf==0.2.5
